Adj log size and direction, fix scope bug
