--- lib/frog.js
@@ -1,9 +1,0 @@
-var Frog = function () {
-  this.x = (canvas.width/2 - frogSize/2);
-  this.y = canvas.height - frogSize;
-  this.width = frogSize;
-  this.height = frogSize;
-};
-
-
-module.exports = Frog;
