+++ lib/index.js
@@ -1,4 +1,5 @@
 var Car = require("./car.js");
+var Frog = require("./frog.js");
 
 var canvas = document.getElementById('game');
 var context = canvas.getContext('2d');
@@ -7,9 +8,9 @@ var frogSize = 25;
 var frogHopDistance = 10;
 
 requestAnimationFrame(function gameLoop() {
+  collisionDetection();
   clearCanvas();
   logCollisionDetection();
-  collisionDetection();
   buildAndStartCars();
   buildAndStartLogs();
   frogger();
@@ -39,6 +40,16 @@ function logCollisionDetection() {
     }
   });
 }
+
+var Frog = function () {
+  // create another object called world
+  // have each object receive world so it knows the confines in interact inside of that world
+
+  this.x = (canvas.width/2 - frogSize/2);
+  this.y = canvas.height - frogSize;
+  this.width = frogSize;
+  this.height = frogSize;
+};
 
 var frog = new Frog();
 
@@ -123,7 +134,7 @@ function Log(x, y, width, height, direction) {
   this.height = height;
   this.direction = direction;
 }
-//could we create prototypes for x and y that increment through Car?
+// could we create prototypes for x and y that increment through Car?
 
 Log.prototype.draw = function() {
   context.fillRect(this.x, this.y, this.width, this.height, this.direction);
@@ -132,19 +143,33 @@ Log.prototype.draw = function() {
 };
 
 Log.prototype.move = function() {
-  if (this.x !== canvas.width) {
-    this.x++;
-  } else {
-    this.x = -this.width;
+
+  if (this.direction === 'right') {
+    if (this.x !== this.canvas.width) {
+      this.x++;
+    } else {
+      this.x = -this.width;
+    }
+    return this;
+
+  } else if (this.direction === 'left') {
+    if (this.x !== -this.width) {
+      this.x--;
+    } else {
+      this.x = this.canvas.width;
+    }
   }
-  return this;
 };
 
 for (var i = 0; i < 5; i++) {
-  x = -50 * i;
-  logs.push(new Log(x, 250, 70, 10));
+  x = -100 * i;
+  logs.push(new Car(x, 250, 70, 30, 'right', canvas));
 }
 
+for (var i = 0; i < 5; i++) {
+  x = 100 * i;
+  logs.push(new Car(x, 150, 70, 30, 'left', canvas));
+}
 
 for (var i = 0; i < 5; i++) {
   x = -200 * i;
